# Add parallel processing to `forc-doc` using rayon to improve performance